Add files via upload
The primary objective of this project is to predict whether a person receives a booking confirmation or not.
The data were prepared by applying cleaning, wrangling and transformation techniques, with important features
extracted to enhance the performance of the model. A classification model was developed using various classification
algorithms to enable accurate predictions. Hyper parameter tuning was conducted to improve the model’s accuracy
and ensure optimal performance.